+++ packages/moonstone/CHANGELOG.md
@@ -13,7 +13,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Marquee.Marquee` to be `moonstone/Marquee.MarqueeBase`
 - `moonstone/VirtualGridList.GridListImageItem` in favor of `moonstone/GridListImageItem`
 - `moonstone/ContextualPopupDecorator` not to remember the last focus
-- `moonstone/ExpandableList` to allow restoring focus to the first selected item when opening
+- `moonstone/ExpandableList` to restore focus to the first selected item after opening
 
 ### Fixed